fix bugs in decodePage and visited set operation
fix a bug when a page response is not a compressed format, add a case
that read the data immediate.
fix a bug, when a url is put into the queue, it will be add into visited
set. In the previous code, just when the url is pop from queue then it
will be added into visited, so maybe the same url maybe visited more
than one time, it's not we want.