add prediction part for textVector model and fixed bugs
1. added prediction part for this model. In prediction phrase, the model
just adjust the vector for text, not the other parameters like word
vectors, softmax parameters.
2. fixed a bug, when we realloc a memory for a point, if the this point
is a argument of a function, the realloc will not change the value of
the point, so I changed to use point reference, by which we can change
point value not only the memory pointed by that point.